Do some refactor work and compile version info into application
